@@ -11,72 +11,84 @@ description: A new Flutter project.
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+1
+version: 1.4.0+8
 
 environment:
-    sdk: ">=2.8.0 <3.0.0"
+    sdk: '>=2.12.0 <3.0.0'
 
 
 dependencies:
-  auto_route:
+  flutter_js: ^0.5.0+3
+  device_preview: ^0.7.4
+  flutter_secure_storage: ^4.2.1
+  flutter_mentions: ^2.0.1
+  photo_view: ^0.12.0
+  flutter_easyloading: ^3.0.0
+  html_editor_enhanced: ^2.3.0
+  flutter_typeahead: ^3.1.1
+  flutter_chips_input: ^1.10.0
+  timelines: ^0.1.0
+  persistent_bottom_nav_bar: ^4.0.2
+  overflow_view: ^0.3.1
   injectable:
-  json_table: ^1.4.0
-  barcode_widget: ^1.5.0
-  shared_preferences: ^0.5.12+2
-  crypto: ^2.1.5
-  device_info: ^1.0.0
-  flutter_local_notifications: ^3.0.0
-  workmanager: ^0.2.3
-  get_it: ^5.0.0
-  fluttertoast: ^7.1.6
-  hive: ^1.4.4
-  hive_flutter: ^0.3.1
-  connectivity: ^2.0.2
-  provider: ^4.1.3
-  open_file: ^3.0.1
-  flutter_svg: ^0.19.0
-  collection: ^1.14.12
-  cached_network_image: ^2.5.0
-  google_fonts: ^1.1.0
-  file_picker: ^2.1.4
-  flutter_form_builder: ^4.2.0
-  flutter_pagewise: ^1.2.3
-  html: ^0.14.0+3
-  url_launcher: ^5.4.7
-  permission_handler: ^5.0.1+1
-  flutter_downloader: ^1.4.4
-  timeago: ^2.0.26
-  path_provider: ^1.6.27
-  flutter_html: ^1.1.1
-  dio: 3.0.9
-  dio_cookie_manager: 1.0.0
-  intl: ^0.16.1
+  json_table: ^2.0.1
+  barcode_widget: ^2.0.1
+  shared_preferences: ^2.0.5
+  crypto: ^3.0.1
+  device_info: ^2.0.0
+  flutter_local_notifications: ^8.2.0
+  # workmanager: ^0.4.0
+  get_it: ^7.2.0
+  fluttertoast: ^8.0.4
+  hive: ^2.0.3
+  hive_flutter: ^1.0.0
+  connectivity: ^3.0.3
+  provider: ^5.0.0
+  open_file: ^3.1.0
+  flutter_svg: ^0.22.0
+  collection: ^1.15.0
+  cached_network_image: ^3.0.0
+  google_fonts: ^2.0.0
+  file_picker: ^4.0.2
+  flutter_form_builder: ^7.0.0
+  form_builder_validators: ^7.1.0
+  flutter_pagewise: ^2.0.1
+  html: ^0.15.0
+  url_launcher: ^6.0.3
+  permission_handler: ^8.1.6
+  flutter_downloader: ^1.6.0
+  timeago: ^3.0.2
+  path_provider: ^2.0.1
+  flutter_html: ^2.0.0
+  dio: ^4.0.0
+  dio_cookie_manager: ^2.0.0
+  intl: ^0.17.0
   flutter:
     sdk: flutter
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
+  cupertino_icons: ^1.0.2
 
 dev_dependencies:
   build_runner:
-  auto_route_generator:
   injectable_generator:
-  hive_generator: ^0.7.1
-  flutter_launcher_icons: "^0.7.3"
+  hive_generator: ^1.0.1
+  flutter_launcher_icons: ^0.9.0
   flutter_test:
     sdk: flutter
-
-dependency_overrides:
-  workmanager:
-    git:
-      url: https://github.com/ened/flutter_workmanager.git
-      ref: 1c4e162ea4e71855856fb9e55a1545b3dae9a25d # flutter 1.20 compatibility
 
 flutter_icons:
   android: "launcher_icon"
   ios: true
   image_path: "assets/frappe_icon.jpg"
+
+dependency_overrides:
+  flutter_colorpicker: ^0.6.0
+  json_table:
+    git:
+      url: https://github.com/sumitbhanushali/json_table
+      ref: 0eee306ab5338340cd63749d3dcfbdd4a5382885
 
 
 # For information on the generic Dart part of this file, see the
