@@ -1,6 +1,14 @@
 import 'package:flutter/material.dart';
 
 import 'package:flutter_form_builder/flutter_form_builder.dart';
+import 'package:form_builder_validators/form_builder_validators.dart';
+
+import 'package:flutter_typeahead/flutter_typeahead.dart';
+import 'package:frappe_app/config/frappe_icons.dart';
+import 'package:frappe_app/config/palette.dart';
+import 'package:frappe_app/model/common.dart';
+import 'package:frappe_app/utils/frappe_icon.dart';
+import 'package:frappe_app/widgets/form_builder_typeahead.dart';
 
 import '../../model/doctype_response.dart';
 
@@ -11,27 +19,27 @@ typedef String SelectionToTextTransformer<T>(T selection);
 
 class AutoComplete extends StatefulWidget {
   final DoctypeField doctypeField;
-  final Map doc;
+  final OnControlChanged? onControlChanged;
 
-  final bool showInputBorder;
-  final bool allowClear;
-  final Function onSuggestionSelected;
-  final Icon prefixIcon;
-  final Color fillColor;
-  final key;
-  final ItemBuilder itemBuilder;
-  final SuggestionsCallback suggestionsCallback;
-  final SelectionToTextTransformer selectionToTextTransformer;
+  final Map? doc;
+  final void Function(dynamic)? onSuggestionSelected;
+  final Widget? suffixIcon;
+  final Key? key;
+  final ItemBuilder? itemBuilder;
+  final SuggestionsCallback? suggestionsCallback;
+  final SelectionToTextTransformer? selectionToTextTransformer;
+  final InputDecoration? inputDecoration;
+  final TextEditingController? controller;
 
   AutoComplete({
-    @required this.doctypeField,
-    @required this.fillColor,
+    required this.doctypeField,
+    this.onControlChanged,
     this.doc,
-    this.prefixIcon,
+    this.controller,
+    this.inputDecoration,
+    this.suffixIcon,
     this.key,
-    this.allowClear = true,
     this.onSuggestionSelected,
-    this.showInputBorder = false,
     this.itemBuilder,
     this.suggestionsCallback,
     this.selectionToTextTransformer,
@@ -43,11 +51,17 @@ class AutoComplete extends StatefulWidget {
 
 class _AutoCompleteState extends State<AutoComplete>
     with Control, ControlInput {
-  final TextEditingController _typeAheadController = TextEditingController();
+  TextEditingController? _typeAheadController;
+
+  @override
+  void initState() {
+    _typeAheadController = widget.controller ?? TextEditingController();
+    super.initState();
+  }
 
   @override
   Widget build(BuildContext context) {
-    List<String Function(dynamic)> validators = [];
+    List<String? Function(dynamic)> validators = [];
 
     var f = setMandatory(widget.doctypeField);
 
@@ -57,71 +71,70 @@ class _AutoCompleteState extends State<AutoComplete>
       );
     }
 
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 12.0),
-      child: Theme(
-        data: Theme.of(context).copyWith(primaryColor: Colors.black),
-        child: FormBuilderTypeAhead(
-          key: widget.key,
-          controller: _typeAheadController,
-          onSuggestionSelected: (item) {
-            if (widget.onSuggestionSelected != null) {
-              widget.onSuggestionSelected(item);
-            }
-          },
-          onChanged: (_) {
-            setState(() {});
-          },
-          validator: FormBuilderValidators.compose(validators),
-          decoration: InputDecoration(
-            filled: true,
-            prefixIcon: widget.prefixIcon,
-            suffixIcon: widget.allowClear
-                ? _typeAheadController.text != ''
-                    ? IconButton(
-                        icon: Icon(Icons.close, color: Colors.black),
-                        onPressed: () {
-                          _typeAheadController.clear();
-                        },
-                      )
-                    : null
-                : null,
-            fillColor: widget.fillColor,
-            enabledBorder: !widget.showInputBorder ? InputBorder.none : null,
-            hintText: widget.doctypeField.label,
-          ),
-          selectionToTextTransformer: widget.selectionToTextTransformer ??
-              (item) {
-                return item.toString();
-              },
-          name: widget.doctypeField.fieldname,
-          itemBuilder: widget.itemBuilder ??
-              (context, item) {
-                return ListTile(
-                  title: Text(
-                    item,
-                  ),
-                );
-              },
-          initialValue: widget.doc != null
-              ? widget.doc[widget.doctypeField.fieldname]
-              : null,
-          suggestionsCallback: widget.suggestionsCallback ??
-              (query) async {
-                var lowercaseQuery = query.toLowerCase();
-                List opts;
-                if (widget.doctypeField.options is String) {
-                  opts = widget.doctypeField.options.split('\n');
-                } else {
-                  opts = widget.doctypeField.options ?? [];
-                }
-                return opts
-                    .where(
-                      (option) => option.toLowerCase().contains(lowercaseQuery),
-                    )
-                    .toList();
-              },
-        ),
+    return Theme(
+      data: Theme.of(context).copyWith(primaryColor: Colors.black),
+      child: FormBuilderTypeAhead(
+        key: widget.key,
+        controller: _typeAheadController,
+        onSuggestionSelected: widget.onSuggestionSelected,
+        onChanged: (val) {
+          if (widget.onControlChanged != null) {
+            widget.onControlChanged!(
+              FieldValue(
+                field: widget.doctypeField,
+                value: val,
+              ),
+            );
+          }
+        },
+        direction: AxisDirection.up,
+        validator: FormBuilderValidators.compose(validators),
+        decoration: widget.inputDecoration ??
+            Palette.formFieldDecoration(
+              suffixIcon: Row(
+                mainAxisSize: MainAxisSize.min,
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  widget.suffixIcon ??
+                      FrappeIcon(
+                        FrappeIcons.select,
+                      ),
+                ],
+              ),
+            ),
+        selectionToTextTransformer: widget.selectionToTextTransformer ??
+            (item) {
+              return item.toString();
+            },
+        name: widget.doctypeField.fieldname,
+        itemBuilder: widget.itemBuilder ??
+            (context, item) {
+              return ListTile(
+                title: Text(
+                  item.toString(),
+                ),
+              );
+            },
+        initialValue: widget.doc != null
+            ? widget.doc![widget.doctypeField.fieldname]
+            : null,
+        suggestionsCallback: widget.suggestionsCallback ??
+            (query) {
+              var lowercaseQuery = query.toLowerCase();
+              List opts;
+              if (widget.doctypeField.options is String) {
+                opts = widget.doctypeField.options.split('\n');
+              } else {
+                opts = widget.doctypeField.options ?? [];
+              }
+              return opts
+                  .where(
+                    (option) => option.toLowerCase().contains(
+                          lowercaseQuery,
+                        ),
+                  )
+                  .toList();
+            },
       ),
     );
   }
