@@ -1,4 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_form_builder/flutter_form_builder.dart';
+import 'package:form_builder_validators/form_builder_validators.dart';
+import 'package:frappe_app/model/common.dart';
 import 'package:frappe_app/model/doctype_response.dart';
 
 import '../../config/frappe_palette.dart';
@@ -9,25 +12,23 @@ import 'base_control.dart';
 import 'base_input.dart';
 
 class Check extends StatelessWidget with Control, ControlInput {
-  final Key key;
   final DoctypeField doctypeField;
-  final Map doc;
-
-  final bool withLabel;
-
-  final Function onChanged;
+  final OnControlChanged? onControlChanged;
+  final Key? key;
+  final Map? doc;
+  final Function? onChanged;
 
   const Check({
-    this.key,
-    @required this.doctypeField,
-    this.doc,
-    this.withLabel,
+    required this.doctypeField,
     this.onChanged,
+    this.onControlChanged,
+    this.key,
+    this.doc,
   });
 
   @override
   Widget build(BuildContext context) {
-    List<String Function(dynamic)> validators = [];
+    List<String? Function(dynamic)> validators = [];
 
     var f = setMandatory(doctypeField);
 
@@ -38,28 +39,39 @@ class Check extends StatelessWidget with Control, ControlInput {
     }
 
     return CustomFormBuilderCheckbox(
+      name: doctypeField.fieldname,
       key: key,
+      enabled:
+          doctypeField.readOnly != null ? doctypeField.readOnly == 0 : true,
       valueTransformer: (val) {
         return val == true ? 1 : 0;
       },
       activeColor: FrappePalette.blue,
-      leadingInput: true,
-      initialValue: doc != null ? doc[doctypeField.fieldname] == 1 : null,
-      onChanged: onChanged != null
-          ? (val) {
-              val = val == true ? 1 : 0;
-              onChanged(val);
-            }
-          : null,
-      attribute: doctypeField.fieldname,
-      label: Text(doctypeField.label),
+      initialValue: doc != null ? doc![doctypeField.fieldname] == 1 : null,
+      onChanged: (val) {
+        if (onControlChanged != null) {
+          onControlChanged!(
+            FieldValue(
+              field: doctypeField,
+              value: val == true ? 1 : 0,
+            ),
+          );
+        }
+      },
+      label: Text(
+        doctypeField.label!,
+        style: TextStyle(
+          color: FrappePalette.grey[700],
+          fontWeight: FontWeight.w400,
+          fontSize: 12,
+        ),
+      ),
       decoration: Palette.formFieldDecoration(
-        withLabel: withLabel,
         label: doctypeField.label,
         filled: false,
         field: "check",
       ),
-      validators: validators,
+      validator: FormBuilderValidators.compose(validators),
     );
   }
 }
