@@ -1,5 +1,12 @@
+import 'dart:convert';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
+import 'package:form_builder_validators/form_builder_validators.dart';
+import 'package:frappe_app/model/offline_storage.dart';
+import 'package:frappe_app/model/system_settings_response.dart';
+import 'package:frappe_app/utils/constants.dart';
+import 'package:intl/intl.dart';
 
 import '../../config/palette.dart';
 import '../../model/doctype_response.dart';
@@ -9,25 +16,20 @@ import 'base_control.dart';
 import 'base_input.dart';
 
 class Date extends StatelessWidget with Control, ControlInput {
-  final Key key;
   final DoctypeField doctypeField;
-  final Map doc;
 
-  final bool withLabel;
-
-  final bool editMode;
+  final Key? key;
+  final Map? doc;
 
   const Date({
     this.key,
-    @required this.doctypeField,
+    required this.doctypeField,
     this.doc,
-    this.withLabel,
-    this.editMode,
   });
 
   @override
   Widget build(BuildContext context) {
-    List<String Function(dynamic)> validators = [];
+    List<String? Function(dynamic)> validators = [];
 
     var f = setMandatory(doctypeField);
 
@@ -37,17 +39,32 @@ class Date extends StatelessWidget with Control, ControlInput {
       );
     }
 
+    var systemSettings = jsonDecode(
+      jsonEncode(
+        OfflineStorage.getItem("systemSettings")["data"],
+      ),
+    );
+
+    var dateFormat = systemSettings != null
+        ? SystemSettingsResponse.fromJson(
+            systemSettings,
+          ).message.defaults.dateFormat
+        : "dd-mm-yyyy";
+
     return FormBuilderDateTimePicker(
       key: key,
       inputType: InputType.date,
       valueTransformer: (val) {
-        return val != null ? val.toIso8601String() : null;
+        return val?.toIso8601String();
       },
-      initialValue: doc != null ? parseDate(doc[doctypeField.fieldname]) : null,
+      format: DateFormat(
+        Constants.frappeFlutterDateFormatMapping[dateFormat],
+      ),
+      initialValue:
+          doc != null ? parseDate(doc![doctypeField.fieldname]) : null,
       keyboardType: TextInputType.number,
       name: doctypeField.fieldname,
       decoration: Palette.formFieldDecoration(
-        withLabel: withLabel,
         label: doctypeField.label,
       ),
       validator: FormBuilderValidators.compose(validators),
