@@ -1,6 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
-import 'package:provider/provider.dart';
+import 'package:flutter_typeahead/flutter_typeahead.dart';
+import 'package:form_builder_validators/form_builder_validators.dart';
+import 'package:frappe_app/config/frappe_icons.dart';
+import 'package:frappe_app/config/palette.dart';
+import 'package:frappe_app/model/common.dart';
+import 'package:frappe_app/utils/frappe_icon.dart';
+import 'package:frappe_app/views/form_view/form_view.dart';
+import 'package:frappe_app/widgets/form_builder_typeahead.dart';
+import 'package:persistent_bottom_nav_bar/persistent-tab-view.dart';
 
 import '../../model/doctype_response.dart';
 import '../../app/locator.dart';
@@ -8,40 +16,38 @@ import '../../services/api/api.dart';
 
 import '../../utils/helpers.dart';
 import '../../model/offline_storage.dart';
-import '../../utils/enums.dart';
 
 import 'base_control.dart';
 import 'base_input.dart';
 
 class LinkField extends StatefulWidget {
   final DoctypeField doctypeField;
-  final Map doc;
+  final Map? doc;
+  final OnControlChanged? onControlChanged;
 
   final key;
-  final bool withLabel;
   final bool showInputBorder;
-  final bool allowClear;
-  final Function onSuggestionSelected;
-  final Icon prefixIcon;
-  final Color fillColor;
-  final ItemBuilder itemBuilder;
-  final SuggestionsCallback suggestionsCallback;
-
-  final List<String Function(dynamic)> validators;
+  final Function? onSuggestionSelected;
+  final Function? noItemsFoundBuilder;
+  final Widget? prefixIcon;
+  final ItemBuilder? itemBuilder;
+  final SuggestionsCallback? suggestionsCallback;
+  final AxisDirection direction;
+  final TextEditingController? controller;
 
   LinkField({
     this.key,
-    @required this.doctypeField,
-    @required this.fillColor,
-    this.withLabel = true,
+    required this.doctypeField,
+    this.onControlChanged,
     this.doc,
     this.prefixIcon,
-    this.allowClear = true,
     this.onSuggestionSelected,
-    this.validators,
+    this.noItemsFoundBuilder,
     this.showInputBorder = false,
     this.itemBuilder,
     this.suggestionsCallback,
+    this.controller,
+    this.direction = AxisDirection.down,
   });
 
   @override
@@ -51,9 +57,9 @@ class LinkField extends StatefulWidget {
 class _LinkFieldState extends State<LinkField> with Control, ControlInput {
   @override
   Widget build(BuildContext context) {
-    List<String Function(dynamic)> validators = [];
-
+    List<String? Function(dynamic)> validators = [];
     var f = setMandatory(widget.doctypeField);
+    late bool enabled;
 
     if (f != null) {
       validators.add(
@@ -61,99 +67,133 @@ class _LinkFieldState extends State<LinkField> with Control, ControlInput {
       );
     }
 
-    var connectionStatus = Provider.of<ConnectivityStatus>(
-      context,
-    );
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 12.0),
-      child: Theme(
-        data: Theme.of(context).copyWith(primaryColor: Colors.black),
-        child: FormBuilderTypeAhead(
-          key: widget.key,
-          onSuggestionSelected: (item) {
-            if (widget.onSuggestionSelected != null) {
-              widget.onSuggestionSelected(item);
-            }
-          },
-          onChanged: (_) {
-            setState(() {});
-          },
-          validator: FormBuilderValidators.compose(validators),
-          decoration: InputDecoration(
-            filled: true,
-            contentPadding: EdgeInsets.symmetric(
-              horizontal: 10,
-            ),
-            border: OutlineInputBorder(
-              borderSide: BorderSide.none,
-              borderRadius: const BorderRadius.all(
-                const Radius.circular(6),
+    // if (widget.doctypeField.readOnly == 1 ||
+    //     (widget.doc != null && widget.doctypeField.setOnlyOnce == 1)) {
+    //   enabled = false;
+    // } else {
+    //   enabled = true;
+    // }
+
+    if (widget.doc != null && widget.doctypeField.setOnlyOnce == 1) {
+      enabled = false;
+    } else {
+      enabled = true;
+    }
+
+    return Theme(
+      data: Theme.of(context).copyWith(primaryColor: Colors.black),
+      child: FormBuilderTypeAhead(
+        key: widget.key,
+        enabled: enabled,
+        controller: widget.controller,
+        initialValue: widget.doc != null
+            ? widget.doc![widget.doctypeField.fieldname]
+            : null,
+        direction: AxisDirection.up,
+        onSuggestionSelected: (item) {
+          var val = item is String
+              ? item
+              : item is Map
+                  ? item["value"]
+                  : null;
+          if (widget.onSuggestionSelected != null) {
+            widget.onSuggestionSelected!(val);
+          }
+          if (widget.onControlChanged != null) {
+            widget.onControlChanged!(
+              FieldValue(
+                field: widget.doctypeField,
+                value: val,
               ),
-            ),
-            prefixIcon: widget.prefixIcon,
-            fillColor: widget.fillColor,
-            hintText: widget.withLabel ? null : widget.doctypeField.label,
-          ),
-          selectionToTextTransformer: (item) {
-            if (item != null) {
-              if (item is Map) {
-                return item["value"];
-              }
+            );
+          }
+        },
+        validator: FormBuilderValidators.compose(validators),
+        decoration: Palette.formFieldDecoration(
+          label: widget.doctypeField.label,
+          suffixIcon: widget.doc?[widget.doctypeField.fieldname] != null &&
+                  widget.doc?[widget.doctypeField.fieldname] != ""
+              ? IconButton(
+                  onPressed: () {
+                    pushNewScreen(
+                      context,
+                      screen: FormView(
+                          doctype: widget.doctypeField.options,
+                          name: widget.doc![widget.doctypeField.fieldname]),
+                    );
+                  },
+                  icon: FrappeIcon(
+                    FrappeIcons.arrow_right_2,
+                    size: 14,
+                  ),
+                )
+              : null,
+        ),
+        selectionToTextTransformer: (item) {
+          if (item != null) {
+            if (item is Map) {
+              return item["value"];
             }
-            return item;
-          },
-          name: widget.doctypeField.fieldname,
-          itemBuilder: widget.itemBuilder ??
-              (context, item) {
+          }
+          return item.toString();
+        },
+        name: widget.doctypeField.fieldname,
+        itemBuilder: widget.itemBuilder ??
+            (context, item) {
+              if (item is Map) {
                 return ListTile(
                   title: Text(
                     item["value"],
                   ),
+                  subtitle: item["description"] != null
+                      ? Text(
+                          item["description"],
+                        )
+                      : null,
                 );
-              },
-          initialValue: widget.doc != null
-              ? widget.doc[widget.doctypeField.fieldname]
-              : null,
-          suggestionsCallback: widget.suggestionsCallback ??
-              (query) async {
-                var lowercaseQuery = query.toLowerCase();
-                var isOnline = await verifyOnline();
-                if ((connectionStatus == null ||
-                        connectionStatus == ConnectivityStatus.offline) &&
-                    !isOnline) {
-                  var linkFull = await OfflineStorage.getItem(
-                      '${widget.doctypeField.options}LinkFull');
-                  linkFull = linkFull["data"];
+              } else {
+                return ListTile(
+                  title: Text(item.toString()),
+                );
+              }
+            },
+        suggestionsCallback: widget.suggestionsCallback ??
+            (query) async {
+              var lowercaseQuery = query.toLowerCase();
+              var isOnline = await verifyOnline();
+              if (!isOnline) {
+                var linkFull = await OfflineStorage.getItem(
+                    '${widget.doctypeField.options}LinkFull');
+                linkFull = linkFull["data"];
 
-                  if (linkFull != null) {
-                    return linkFull["results"].where(
-                      (link) {
-                        return (link["value"] as String)
-                            .toLowerCase()
-                            .contains(lowercaseQuery);
-                      },
-                    ).toList();
-                  } else {
-                    var queryLink = await OfflineStorage.getItem(
-                        '$lowercaseQuery${widget.doctypeField.options}Link');
-                    queryLink = queryLink["data"];
-
-                    if (queryLink != null) {
-                      return queryLink["results"];
-                    } else {
-                      return [];
-                    }
-                  }
+                if (linkFull != null) {
+                  return linkFull["results"].where(
+                    (link) {
+                      return (link["value"] as String)
+                          .toLowerCase()
+                          .contains(lowercaseQuery);
+                    },
+                  ).toList();
                 } else {
-                  var response = await locator<Api>().searchLink(
-                    doctype: widget.doctypeField.options,
-                    txt: lowercaseQuery,
-                  );
+                  var queryLink = await OfflineStorage.getItem(
+                      '$lowercaseQuery${widget.doctypeField.options}Link');
+                  queryLink = queryLink["data"];
 
-                  return response["results"];
+                  if (queryLink != null) {
+                    return queryLink["results"];
+                  } else {
+                    return [];
+                  }
                 }
-              },
-        ),
+              } else {
+                var response = await locator<Api>().searchLink(
+                  doctype: widget.doctypeField.options,
+                  txt: lowercaseQuery,
+                );
+
+                return response["results"];
+              }
+            },
       ),
     );
   }
