@@ -1,5 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
+import 'package:form_builder_validators/form_builder_validators.dart';
+import 'package:frappe_app/config/frappe_icons.dart';
+import 'package:frappe_app/model/common.dart';
+import 'package:frappe_app/utils/frappe_icon.dart';
 
 import '../../config/palette.dart';
 import '../../model/doctype_response.dart';
@@ -8,25 +12,22 @@ import 'base_control.dart';
 import 'base_input.dart';
 
 class Select extends StatelessWidget with Control, ControlInput {
-  final Key key;
   final DoctypeField doctypeField;
-  final Map doc;
+  final OnControlChanged? onControlChanged;
 
-  final bool allowClear;
-
-  final bool withLabel;
+  final Key? key;
+  final Map? doc;
 
   const Select({
     this.key,
-    this.doctypeField,
+    required this.doctypeField,
     this.doc,
-    this.allowClear,
-    this.withLabel,
+    this.onControlChanged,
   });
 
   @override
   Widget build(BuildContext context) {
-    List<String Function(dynamic)> validators = [];
+    List<String? Function(dynamic)> validators = [];
 
     var f = setMandatory(doctypeField);
 
@@ -45,13 +46,25 @@ class Select extends StatelessWidget with Control, ControlInput {
 
     return FormBuilderDropdown(
       key: key,
-      initialValue:
-          doc != null ? doc[doctypeField.fieldname] : doctypeField.defaultValue,
-      allowClear: allowClear,
+      onChanged: (dynamic val) {
+        if (onControlChanged != null) {
+          onControlChanged!(
+            FieldValue(
+              field: doctypeField,
+              value: val,
+            ),
+          );
+        }
+      },
+      icon: FrappeIcon(
+        FrappeIcons.select,
+      ),
+      initialValue: doc != null
+          ? doc![doctypeField.fieldname]
+          : doctypeField.defaultValue,
       name: doctypeField.fieldname,
-      hint: !withLabel ? Text(doctypeField.label) : null,
+      hint: Text(doctypeField.label!),
       decoration: Palette.formFieldDecoration(
-        withLabel: withLabel,
         label: doctypeField.label,
       ),
       validator: FormBuilderValidators.compose(validators),
