@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
+import 'package:form_builder_validators/form_builder_validators.dart';
 import 'package:frappe_app/config/palette.dart';
 import 'package:frappe_app/model/doctype_response.dart';
 
@@ -7,22 +8,22 @@ import 'base_control.dart';
 import 'base_input.dart';
 
 class SmallText extends StatelessWidget with Control, ControlInput {
-  final Key key;
   final DoctypeField doctypeField;
-  final Map doc;
+  final void Function(String?)? onChanged;
 
-  final bool withLabel;
+  final Key? key;
+  final Map? doc;
 
   const SmallText({
+    required this.doctypeField,
+    this.onChanged,
     this.key,
-    @required this.doctypeField,
     this.doc,
-    this.withLabel,
   });
 
   @override
   Widget build(BuildContext context) {
-    List<String Function(dynamic)> validators = [];
+    List<String? Function(dynamic)> validators = [];
 
     var f = setMandatory(doctypeField);
 
@@ -34,10 +35,11 @@ class SmallText extends StatelessWidget with Control, ControlInput {
 
     return FormBuilderTextField(
       key: key,
-      initialValue: doc != null ? doc[doctypeField.fieldname] : null,
+      onChanged: onChanged,
+      readOnly: doctypeField.readOnly == 1 ? true : false,
+      initialValue: doc != null ? doc![doctypeField.fieldname] : null,
       name: doctypeField.fieldname,
       decoration: Palette.formFieldDecoration(
-        withLabel: withLabel,
         label: doctypeField.label,
       ),
       validator: FormBuilderValidators.compose(validators),
