@@ -1,29 +1,31 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
+import 'package:form_builder_validators/form_builder_validators.dart';
+import 'package:frappe_app/widgets/form_builder_text_editor.dart';
 
-import '../../config/palette.dart';
 import '../../model/doctype_response.dart';
 
 import 'base_control.dart';
 import 'base_input.dart';
 
 class TextEditor extends StatelessWidget with Control, ControlInput {
-  final Key key;
   final DoctypeField doctypeField;
-  final Map doc;
-
-  final bool withLabel;
+  final Key? key;
+  final Map? doc;
+  final Color? color;
+  final bool fullHeight;
 
   const TextEditor({
+    required this.doctypeField,
     this.key,
-    @required this.doctypeField,
     this.doc,
-    this.withLabel,
+    this.color,
+    this.fullHeight = false,
   });
 
   @override
   Widget build(BuildContext context) {
-    List<String Function(dynamic)> validators = [];
+    List<String? Function(dynamic)> validators = [];
 
     var f = setMandatory(doctypeField);
 
@@ -32,15 +34,14 @@ class TextEditor extends StatelessWidget with Control, ControlInput {
         f(context),
       );
     }
-    return FormBuilderTextField(
+
+    return FormBuilderTextEditor(
       key: key,
-      maxLines: 10,
-      initialValue: doc != null ? doc[doctypeField.fieldname] : null,
+      fullHeight: fullHeight,
+      initialValue: doc != null ? doc![doctypeField.fieldname] : null,
       name: doctypeField.fieldname,
-      decoration: Palette.formFieldDecoration(
-        withLabel: withLabel,
-        label: doctypeField.label,
-      ),
+      context: context,
+      color: color,
       validator: FormBuilderValidators.compose(validators),
     );
   }
