@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
+import 'package:form_builder_validators/form_builder_validators.dart';
 import 'package:intl/intl.dart';
 
 import 'package:frappe_app/config/palette.dart';
@@ -9,23 +10,19 @@ import 'base_control.dart';
 import 'base_input.dart';
 
 class Time extends StatelessWidget with Control, ControlInput {
-  final Key key;
   final DoctypeField doctypeField;
-  final Map doc;
-  final bool withLabel;
-  final bool editMode;
+  final Key? key;
+  final Map? doc;
 
   const Time({
+    required this.doctypeField,
     this.key,
-    @required this.doctypeField,
     this.doc,
-    this.withLabel,
-    this.editMode,
   });
 
   @override
   Widget build(BuildContext context) {
-    List<String Function(dynamic)> validators = [];
+    List<String? Function(dynamic?)> validators = [];
 
     var f = setMandatory(doctypeField);
 
@@ -35,21 +32,31 @@ class Time extends StatelessWidget with Control, ControlInput {
       );
     }
 
+    var initialValue;
+
+    if (doc != null) {
+      var value = doc![doctypeField.fieldname];
+
+      if (value != null) {
+        if ((value as String).contains("T")) {
+          value = doc![doctypeField.fieldname].split("T")[1];
+        }
+        initialValue = DateFormat.Hms().parse(
+          value,
+        );
+      }
+    }
+
     return FormBuilderDateTimePicker(
       key: key,
-      initialValue: doc != null
-          ? DateFormat.Hms().parse(
-              doc[doctypeField.fieldname],
-            )
-          : null,
+      initialValue: initialValue,
       inputType: InputType.time,
       valueTransformer: (val) {
-        return val != null ? val.toIso8601String() : null;
+        return val?.toIso8601String();
       },
       keyboardType: TextInputType.number,
       name: doctypeField.fieldname,
       decoration: Palette.formFieldDecoration(
-        withLabel: withLabel,
         label: doctypeField.label,
       ),
       validator: FormBuilderValidators.compose(validators),
