@@ -1,11 +1,11 @@
 class DeskSidebarItemsResponse {
-  List<DeskMessage> message;
+  late List<DeskMessage> message;
 
-  DeskSidebarItemsResponse({this.message});
+  DeskSidebarItemsResponse({required this.message});
 
   DeskSidebarItemsResponse.fromJson(Map<dynamic, dynamic> json) {
     if (json['message'] != null) {
-      message = new List<DeskMessage>();
+      message = [];
       json['message'].forEach((v) {
         message.add(new DeskMessage.fromJson(v));
       });
@@ -14,21 +14,27 @@ class DeskSidebarItemsResponse {
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
-    if (this.message != null) {
-      data['message'] = this.message.map((v) => v.toJson()).toList();
-    }
+    data['message'] = this.message.map((v) => v.toJson()).toList();
     return data;
   }
 }
 
 class DeskMessage {
-  String name;
-  String category;
-  String icon;
-  String module;
-  String label;
+  late String name;
+  late String? category;
+  late String? icon;
+  late String module;
+  late String label;
+  late String? content;
 
-  DeskMessage({this.name, this.category, this.icon, this.module, this.label});
+  DeskMessage({
+    required this.name,
+    this.category,
+    this.icon,
+    this.content,
+    required this.module,
+    required this.label,
+  });
 
   DeskMessage.fromJson(Map<dynamic, dynamic> json) {
     name = json['name'];
@@ -36,6 +42,7 @@ class DeskMessage {
     icon = json['icon'];
     module = json['module'];
     label = json['label'];
+    content = json['content'];
   }
 
   Map<String, dynamic> toJson() {
@@ -45,6 +52,7 @@ class DeskMessage {
     data['icon'] = this.icon;
     data['module'] = this.module;
     data['label'] = this.label;
+    data['content'] = this.content;
     return data;
   }
 }
