@@ -1,73 +1,63 @@
 class DesktopPageResponse {
-  DesktopPageMessage message;
+  late DesktopPageMessage message;
 
-  DesktopPageResponse({this.message});
+  DesktopPageResponse({required this.message});
 
   DesktopPageResponse.fromJson(Map<dynamic, dynamic> json) {
-    message = json['message'] != null
-        ? new DesktopPageMessage.fromJson(json['message'])
-        : null;
+    message = DesktopPageMessage.fromJson(json['message']);
   }
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
-    if (this.message != null) {
-      data['message'] = this.message.toJson();
-    }
+    data['message'] = this.message.toJson();
     return data;
   }
 }
 
 class DesktopPageMessage {
-  DesktopPageCharts charts;
-  DesktopPageShortcuts shortcuts;
-  DesktopPageCards cards;
-  bool allowCustomization;
+  late DesktopPageCharts charts;
+  late DesktopPageShortcuts shortcuts;
+  late DesktopPageCards cards;
+  late bool allowCustomization;
 
-  DesktopPageMessage(
-      {this.charts, this.shortcuts, this.cards, this.allowCustomization});
+  DesktopPageMessage({
+    required this.charts,
+    required this.shortcuts,
+    required this.cards,
+    required this.allowCustomization,
+  });
 
   DesktopPageMessage.fromJson(Map<dynamic, dynamic> json) {
-    charts = json['charts'] != null
-        ? new DesktopPageCharts.fromJson(json['charts'])
-        : null;
-    shortcuts = json['shortcuts'] != null
-        ? new DesktopPageShortcuts.fromJson(json['shortcuts'])
-        : null;
-    cards = json['cards'] != null
-        ? new DesktopPageCards.fromJson(json['cards'])
-        : null;
+    charts = DesktopPageCharts.fromJson(json['charts']);
+
+    shortcuts = DesktopPageShortcuts.fromJson(json['shortcuts']);
+    cards = DesktopPageCards.fromJson(json['cards']);
+
     allowCustomization = json['allow_customization'];
   }
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
-    if (this.charts != null) {
-      data['charts'] = this.charts.toJson();
-    }
-    if (this.shortcuts != null) {
-      data['shortcuts'] = this.shortcuts.toJson();
-    }
-    if (this.cards != null) {
-      data['cards'] = this.cards.toJson();
-    }
+    data['charts'] = this.charts.toJson();
+    data['shortcuts'] = this.shortcuts.toJson();
+    data['cards'] = this.cards.toJson();
     data['allow_customization'] = this.allowCustomization;
     return data;
   }
 }
 
 class DesktopPageCharts {
-  String label;
-  List<ChartItem> items;
+  late String? label;
+  late List<ChartItem>? items;
 
   DesktopPageCharts({this.label, this.items});
 
   DesktopPageCharts.fromJson(Map<dynamic, dynamic> json) {
     label = json['label'];
     if (json['items'] != null) {
-      items = new List<ChartItem>();
+      items = [];
       json['items'].forEach((v) {
-        items.add(new ChartItem.fromJson(v));
+        items?.add(new ChartItem.fromJson(v));
       });
     }
   }
@@ -76,22 +66,25 @@ class DesktopPageCharts {
     final Map<String, dynamic> data = new Map<String, dynamic>();
     data['label'] = this.label;
     if (this.items != null) {
-      data['items'] = this.items.map((v) => v.toJson()).toList();
+      data['items'] = this.items?.map((v) => v.toJson()).toList();
     }
     return data;
   }
 }
 
 class DesktopPageShortcuts {
-  String label;
-  List<ShortcutItem> items;
+  late String label;
+  late List<ShortcutItem> items;
 
-  DesktopPageShortcuts({this.label, this.items});
+  DesktopPageShortcuts({
+    required this.label,
+    required this.items,
+  });
 
   DesktopPageShortcuts.fromJson(Map<dynamic, dynamic> json) {
     label = json['label'];
     if (json['items'] != null) {
-      items = new List<ShortcutItem>();
+      items = [];
       json['items'].forEach((v) {
         items.add(new ShortcutItem.fromJson(v));
       });
@@ -101,25 +94,26 @@ class DesktopPageShortcuts {
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
     data['label'] = this.label;
-    if (this.items != null) {
-      data['items'] = this.items.map((v) => v.toJson()).toList();
-    }
+    data['items'] = this.items.map((v) => v.toJson()).toList();
     return data;
   }
 }
 
 class DesktopPageCards {
-  String label;
-  List<CardItem> items;
+  late String label;
+  late List<CardItem> items;
 
-  DesktopPageCards({this.label, this.items});
+  DesktopPageCards({
+    required this.label,
+    required this.items,
+  });
 
   DesktopPageCards.fromJson(Map<dynamic, dynamic> json) {
     label = json['label'];
     if (json['items'] != null) {
-      items = new List<CardItem>();
+      items = [];
       json['items'].forEach((v) {
-        items.add(new CardItem.fromJson(v));
+        items.add(CardItem.fromJson(v));
       });
     }
   }
@@ -127,27 +121,25 @@ class DesktopPageCards {
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
     data['label'] = this.label;
-    if (this.items != null) {
-      data['items'] = this.items.map((v) => v.toJson()).toList();
-    }
+    data['items'] = this.items.map((v) => v.toJson()).toList();
     return data;
   }
 }
 
 class ChartItem {
-  String name;
-  String owner;
-  String creation;
-  String modified;
-  String modifiedBy;
-  String parent;
-  String parentfield;
-  String parenttype;
-  int idx;
-  int docstatus;
-  String chartName;
-  String label;
-  String doctype;
+  late String? name;
+  late String? owner;
+  late String? creation;
+  late String? modified;
+  late String? modifiedBy;
+  late String? parent;
+  late String? parentfield;
+  late String? parenttype;
+  late int? idx;
+  late int? docstatus;
+  late String? chartName;
+  late String? label;
+  late String? doctype;
 
   ChartItem(
       {this.name,
@@ -200,27 +192,27 @@ class ChartItem {
 }
 
 class ShortcutItem {
-  String name;
-  String owner;
-  String creation;
-  String modified;
-  String modifiedBy;
-  String parent;
-  String parentfield;
-  String parenttype;
-  int idx;
-  int docstatus;
-  String type;
-  String linkTo;
-  String docView;
-  String label;
-  dynamic icon;
-  dynamic restrictToDomain;
-  String statsFilter;
-  String color;
-  String format;
-  String doctype;
-  int isQueryReport;
+  late String? name;
+  late String? owner;
+  late String? creation;
+  late String? modified;
+  late String? modifiedBy;
+  late String? parent;
+  late String? parentfield;
+  late String? parenttype;
+  late int? idx;
+  late int? docstatus;
+  late String? type;
+  late String linkTo;
+  late String? docView;
+  late String label;
+  late dynamic? icon;
+  late dynamic? restrictToDomain;
+  late String? statsFilter;
+  late String? color;
+  late String? format;
+  late String? doctype;
+  late int? isQueryReport;
 
   ShortcutItem(
       {this.name,
@@ -234,9 +226,9 @@ class ShortcutItem {
       this.idx,
       this.docstatus,
       this.type,
-      this.linkTo,
+      required this.linkTo,
       this.docView,
-      this.label,
+      required this.label,
       this.icon,
       this.restrictToDomain,
       this.statsFilter,
@@ -297,20 +289,20 @@ class ShortcutItem {
 }
 
 class CardItem {
-  String name;
-  String owner;
-  String creation;
-  String modified;
-  String modifiedBy;
-  String parent;
-  String parentfield;
-  String parenttype;
-  int idx;
-  int docstatus;
-  String label;
-  int hidden;
-  List<CardItemLink> links;
-  String doctype;
+  late String? name;
+  late String? owner;
+  late String? creation;
+  late String? modified;
+  late String? modifiedBy;
+  late String? parent;
+  late String? parentfield;
+  late String? parenttype;
+  late int? idx;
+  late int? docstatus;
+  late String label;
+  late dynamic hidden;
+  late List<CardItemLink> links;
+  late String? doctype;
 
   CardItem(
       {this.name,
@@ -323,9 +315,9 @@ class CardItem {
       this.parenttype,
       this.idx,
       this.docstatus,
-      this.label,
+      required this.label,
       this.hidden,
-      this.links,
+      required this.links,
       this.doctype});
 
   CardItem.fromJson(Map<dynamic, dynamic> json) {
@@ -342,9 +334,9 @@ class CardItem {
     label = json['label'];
     hidden = json['hidden'];
     if (json['links'] != null) {
-      links = new List<CardItemLink>();
+      links = [];
       json['links'].forEach((v) {
-        links.add(new CardItemLink.fromJson(v));
+        links.add(CardItemLink.fromJson(v));
       });
     }
     doctype = json['doctype'];
@@ -364,32 +356,30 @@ class CardItem {
     data['docstatus'] = this.docstatus;
     data['label'] = this.label;
     data['hidden'] = this.hidden;
-    if (this.links != null) {
-      data['links'] = this.links.map((v) => v.toJson()).toList();
-    }
+    data['links'] = this.links.map((v) => v.toJson()).toList();
     data['doctype'] = this.doctype;
     return data;
   }
 }
 
 class CardItemLink {
-  String description;
-  String label;
-  String name;
-  int onboard;
-  String type;
-  dynamic count;
-  dynamic dependencies;
-  String doctype;
-  dynamic isQueryReport;
-  dynamic incompleteDependencies;
-  String icon;
-  String linkTo;
+  late String? description;
+  late String label;
+  late String name;
+  late int? onboard;
+  late String? type;
+  late dynamic? count;
+  late dynamic? dependencies;
+  late String? doctype;
+  late dynamic? isQueryReport;
+  late dynamic? incompleteDependencies;
+  late String? icon;
+  late String? linkTo;
 
   CardItemLink(
       {this.description,
-      this.label,
-      this.name,
+      required this.label,
+      required this.name,
       this.onboard,
       this.type,
       this.count,
