@@ -3,8 +3,8 @@ import 'package:flutter_svg/flutter_svg.dart';
 
 class FrappeIcon extends StatelessWidget {
   final String path;
-  final Color color;
-  final double size;
+  final Color? color;
+  final double? size;
 
   const FrappeIcon(
     this.path, {
