@@ -1,146 +1,348 @@
+import 'dart:convert';
+import 'dart:io';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_html/flutter_html.dart';
-import 'package:frappe_app/app/locator.dart';
-import 'package:frappe_app/app/router.gr.dart';
-import 'package:frappe_app/services/navigation_service.dart';
+import 'package:frappe_app/config/frappe_icons.dart';
+import 'package:frappe_app/config/frappe_palette.dart';
+import 'package:frappe_app/model/get_doc_response.dart';
+import 'package:frappe_app/utils/dio_helper.dart';
+import 'package:frappe_app/utils/frappe_icon.dart';
+import 'package:frappe_app/utils/helpers.dart';
+import 'package:frappe_app/views/login/login_view.dart';
+import 'package:open_file/open_file.dart';
 
-import 'package:html/parser.dart';
 import 'package:timeago/timeago.dart' as timeago;
-
-import '../config/palette.dart';
+import 'package:url_launcher/url_launcher.dart';
 
 import '../model/config.dart';
 
 import '../widgets/user_avatar.dart';
+import 'package:photo_view/photo_view.dart';
 
-class EmailBox extends StatelessWidget {
-  final Map data;
+class EmailBox extends StatefulWidget {
+  final Communication data;
+  final Function onReplyTo;
+  final Function onReplyAll;
 
-  EmailBox(this.data);
-
-  @override
-  Widget build(BuildContext context) {
-    var time = timeago.format(DateTime.parse(data["creation"]));
-
-    var document = parse(data["content"]);
-    String parsedContent = parse(document.body.text).documentElement.text;
-
-    return Card(
-      elevation: 0,
-      child: Padding(
-        padding: const EdgeInsets.symmetric(vertical: 8.0),
-        child: ListTile(
-          onTap: () {
-            locator<NavigationService>().navigateTo(
-              Routes.viewEmail,
-              arguments: ViewEmailArguments(
-                time: time,
-                title: data["subject"],
-                senderFullName: data["sender_full_name"],
-                sender: data["sender"],
-                content: data["content"],
-              ),
-            );
-          },
-          subtitle:
-              Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
-            Container(
-              padding: const EdgeInsets.only(bottom: 8.0, top: 4.0),
-              child: Text(
-                data["subject"],
-                maxLines: 1,
-                overflow: TextOverflow.ellipsis,
-                style: TextStyle(color: Colors.black),
-              ),
-            ),
-            Text(
-              parsedContent,
-              maxLines: 1,
-              overflow: TextOverflow.ellipsis,
-            )
-          ]),
-          title: Row(
-            children: [
-              Text('${data["sender_full_name"]}'),
-              Spacer(),
-              Text(
-                time,
-                style: TextStyle(
-                  fontSize: 12,
-                ),
-              )
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class ViewEmail extends StatelessWidget {
-  final String title;
-  final String time;
-  final String senderFullName;
-  final String sender;
-  final String content;
-
-  ViewEmail({
-    @required this.title,
-    @required this.time,
-    @required this.senderFullName,
-    @required this.sender,
-    @required this.content,
+  EmailBox({
+    required this.data,
+    required this.onReplyAll,
+    required this.onReplyTo,
   });
 
   @override
+  _EmailBoxState createState() => _EmailBoxState();
+}
+
+class _EmailBoxState extends State<EmailBox> {
+  late bool _isExpanded;
+
+  @override
+  void initState() {
+    _isExpanded = false;
+    super.initState();
+  }
+
+  @override
   Widget build(BuildContext context) {
-    var document = parse(content);
-    var imgs = document.getElementsByTagName('img');
+    var message = widget.data.content;
 
-    imgs.forEach((img) {
-      if (Uri.parse(img.attributes["src"]).hasAbsolutePath) {
-        img.attributes["src"] = "${Config().baseUrl}${img.attributes["src"]}";
-      }
-    });
+    var time = timeago.format(DateTime.parse(widget.data.creation));
+    var attachments = (json.decode(widget.data.attachments) as List)
+        .map(
+          (attachment) => Attachments.fromJson(
+            attachment,
+          ),
+        )
+        .toList();
 
-    return Scaffold(
-      appBar: AppBar(),
-      body: SingleChildScrollView(
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: <Widget>[
-            Card(
-              elevation: 4,
-              child: Container(
-                color: Palette.bgColor,
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: <Widget>[
-                    Padding(
-                      padding: const EdgeInsets.all(8.0),
-                      child: Text(
-                        title,
-                        style: TextStyle(
-                            fontWeight: FontWeight.bold, fontSize: 18),
-                      ),
-                    ),
-                    ListTile(
-                      leading: UserAvatar(
-                        uid: sender,
-                      ),
-                      title: Text(senderFullName),
-                      subtitle: Text(time),
-                    ),
-                  ],
-                ),
-              ),
+    return Padding(
+      padding: const EdgeInsets.only(left: 3.0),
+      child: GestureDetector(
+        onTap: () {
+          setState(
+            () {
+              _isExpanded = !_isExpanded;
+            },
+          );
+        },
+        child: ConstrainedBox(
+          constraints: BoxConstraints(
+            maxHeight: _isExpanded ? double.infinity : 184,
+          ),
+          child: Card(
+            color: Colors.white,
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(8),
             ),
-            Container(
-              color: Colors.white,
-              padding: const EdgeInsets.all(8.0),
-              child: Html(data: document.outerHtml),
-            )
-          ],
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                ListTile(
+                  leading: UserAvatar(
+                    uid: widget.data.sender,
+                  ),
+                  title: Row(
+                    children: [
+                      Flexible(
+                        child: Text(
+                          widget.data.senderFullName,
+                        ),
+                      ),
+                      SizedBox(
+                        width: 10,
+                      ),
+                      if (widget.data.deliveryStatus != "")
+                        widget.data.deliveryStatus == "Sent"
+                            ? FrappeIcon(
+                                FrappeIcons.unread_status,
+                              )
+                            : FrappeIcon(
+                                FrappeIcons.read_status,
+                                color: FrappePalette.blue,
+                              ),
+                    ],
+                  ),
+                  subtitle: Text(
+                    time,
+                  ),
+                  trailing: Row(
+                    mainAxisSize: MainAxisSize.min,
+                    children: [
+                      // FrappeIcon(
+                      //   FrappeIcons.favourite_resting,
+                      //   color: FrappePalette.grey,
+                      //   size: 20,
+                      // ),
+                      // SizedBox(
+                      //   width: 11,
+                      // ),
+                      GestureDetector(
+                        onTap: () {
+                          widget.onReplyTo();
+                        },
+                        child: FrappeIcon(
+                          FrappeIcons.reply_alt,
+                          color: FrappePalette.grey,
+                          size: 18,
+                        ),
+                      ),
+                      SizedBox(
+                        width: 11,
+                      ),
+                      GestureDetector(
+                        onTap: () {
+                          widget.onReplyAll();
+                        },
+                        child: FrappeIcon(
+                          FrappeIcons.reply_all,
+                        ),
+                      ),
+                      // SizedBox(
+                      //   width: 11,
+                      // ),
+                      // FrappeIcon(
+                      //   FrappeIcons.dot_vertical,
+                      // ),
+                    ],
+                  ),
+                ),
+                if (attachments.isNotEmpty && _isExpanded)
+                  Container(
+                    height: 50,
+                    child: ListView.builder(
+                      shrinkWrap: true,
+                      itemCount: attachments.length,
+                      scrollDirection: Axis.horizontal,
+                      itemBuilder: (context, index) {
+                        var fileName =
+                            attachments[index].fileUrl.split("/").last;
+                        var extension = fileName.split('.').last;
+
+                        return InkWell(
+                          onTap: () async {
+                            var downloadPath = await getDownloadPath();
+
+                            var filePath = "$downloadPath$fileName";
+
+                            var fileExists = await File(filePath).exists();
+
+                            if (fileExists) {
+                              OpenFile.open(filePath);
+                            } else {
+                              downloadFile(
+                                  attachments[index].fileUrl, downloadPath);
+                            }
+                          },
+                          child: Padding(
+                            padding: EdgeInsets.symmetric(
+                              horizontal: 12,
+                            ),
+                            child: Container(
+                              width: 162,
+                              decoration: BoxDecoration(
+                                borderRadius: BorderRadius.circular(8),
+                                border:
+                                    Border.all(color: FrappePalette.grey[200]!),
+                              ),
+                              child: Row(
+                                children: [
+                                  SizedBox(
+                                    width: 10,
+                                  ),
+                                  CircleAvatar(
+                                    backgroundColor: FrappePalette.grey[100],
+                                    child: FrappeIcon(
+                                      FrappeIcons.small_file,
+                                    ),
+                                  ),
+                                  SizedBox(
+                                    width: 8,
+                                  ),
+                                  Flexible(
+                                    child: Column(
+                                      mainAxisAlignment:
+                                          MainAxisAlignment.center,
+                                      crossAxisAlignment:
+                                          CrossAxisAlignment.start,
+                                      children: [
+                                        Text(
+                                          fileName,
+                                          overflow: TextOverflow.ellipsis,
+                                          style: TextStyle(
+                                            color: FrappePalette.grey[900],
+                                            fontSize: 13,
+                                          ),
+                                        ),
+                                        Text(
+                                          extension.toUpperCase(),
+                                          style: TextStyle(
+                                            color: FrappePalette.grey[600],
+                                            fontWeight: FontWeight.w400,
+                                            fontSize: 12,
+                                          ),
+                                        ),
+                                      ],
+                                    ),
+                                  ),
+                                  SizedBox(
+                                    width: 10,
+                                  ),
+                                ],
+                              ),
+                            ),
+                          ),
+                        );
+                      },
+                    ),
+                  ),
+                Flexible(
+                  child: ShaderMask(
+                    shaderCallback: (rect) {
+                      return LinearGradient(
+                        begin: Alignment(0.0, 0.6),
+                        end: Alignment.bottomCenter,
+                        colors: [
+                          Colors.white,
+                          _isExpanded ? Colors.white : Colors.white10
+                        ],
+                      ).createShader(
+                        Rect.fromLTRB(
+                          0,
+                          0,
+                          rect.width,
+                          rect.height,
+                        ),
+                      );
+                    },
+                    child: Container(
+                      color: Colors.white,
+                      padding: const EdgeInsets.symmetric(horizontal: 8.0),
+                      child: Html(
+                        data: message,
+                        customRender: {
+                          "img": (renderContext, child) {
+                            var src = renderContext.tree.attributes['src'];
+                            if (src != null) {
+                              if (!src.startsWith("http")) {
+                                src = Config().baseUrl! + src;
+                              }
+                              return GestureDetector(
+                                onTap: () => Navigator.of(
+                                  context,
+                                  rootNavigator: true,
+                                ).push(
+                                  MaterialPageRoute(
+                                    builder: (context) => Scaffold(
+                                      appBar: AppBar(
+                                        elevation: 0.8,
+                                      ),
+                                      body: PhotoView(
+                                        imageProvider: NetworkImage(
+                                          src!,
+                                        ),
+                                      ),
+                                    ),
+                                  ),
+                                ),
+                                child: Image.network(
+                                  src,
+                                  headers: {
+                                    HttpHeaders.cookieHeader:
+                                        DioHelper.cookies!,
+                                  },
+                                ),
+                              );
+                            }
+                          },
+                        },
+                        customImageRenders: {
+                          networkSourceMatcher(domains: [
+                            Config().baseUrl!,
+                          ]): networkImageRender(
+                            headers: {
+                              HttpHeaders.cookieHeader: DioHelper.cookies!,
+                            },
+                            altWidget: (alt) => Text(alt ?? ""),
+                            loadingWidget: () => Text("Loading..."),
+                          ),
+                          // for relative paths, prefix with a base url
+                          (attr, _) =>
+                                  attr["src"] != null &&
+                                  !attr["src"]!.startsWith("http"):
+                              networkImageRender(
+                            headers: {
+                              HttpHeaders.cookieHeader: DioHelper.cookies!,
+                            },
+                            mapUrl: (url) => Config().baseUrl! + url!,
+                          ),
+                          // Custom placeholder image for broken links
+                          networkSourceMatcher(): networkImageRender(
+                              altWidget: (_) => FrappeLogo()),
+                        },
+                        onLinkTap: (url, _, __, ___) async {
+                          print("Opening $url...");
+                          if (url != null) {
+                            if (await canLaunch(url)) {
+                              await launch(
+                                url,
+                                headers: {
+                                  HttpHeaders.cookieHeader: DioHelper.cookies!,
+                                },
+                              );
+                            } else {
+                              throw 'Could not launch $url';
+                            }
+                          }
+                        },
+                      ),
+                    ),
+                  ),
+                )
+              ],
+            ),
+          ),
         ),
       ),
     );
