@@ -6,15 +6,20 @@ import '../config/palette.dart';
 import '../services/api/api.dart';
 import '../utils/frappe_icon.dart';
 
+// ignore: must_be_immutable
 class LikeDoc extends StatefulWidget {
   bool isFav;
   final String doctype;
   final String name;
+  final Color? iconColor;
+  final Function? successCallback;
 
   LikeDoc({
-    @required this.isFav,
-    @required this.doctype,
-    @required this.name,
+    required this.isFav,
+    required this.doctype,
+    required this.name,
+    this.successCallback,
+    this.iconColor,
   });
 
   @override
@@ -34,6 +39,9 @@ class _LikeDocState extends State<LikeDoc> {
     );
 
     if (response.statusCode == 200) {
+      if (widget.successCallback != null) {
+        widget.successCallback!();
+      }
       return;
     } else {
       setState(() {
@@ -52,7 +60,7 @@ class _LikeDocState extends State<LikeDoc> {
               ? FrappeIcons.favourite_active
               : FrappeIcons.favourite_resting,
           size: 18,
-          color: widget.isFav ? null : Palette.iconColor,
+          color: widget.iconColor ?? Palette.iconColor,
         ),
       ),
       onTap: _toggleFav,
