@@ -1,49 +1,96 @@
-import 'dart:io';
+import 'dart:math';
 
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
-
-import '../config/palette.dart';
+import 'package:frappe_app/config/frappe_palette.dart';
+import 'package:frappe_app/utils/enums.dart';
 
 import '../model/offline_storage.dart';
-import '../utils/dio_helper.dart';
 import '../utils/helpers.dart';
 import '../utils/http.dart';
 
 class UserAvatar extends StatelessWidget {
   final String uid;
-  final double size;
+  final double? size;
+  final ImageShape shape;
 
   UserAvatar({
-    this.uid,
+    required this.uid,
+    this.shape = ImageShape.circle,
     this.size,
   });
 
-  static Widget renderShape({
-    String txt,
-    ImageProvider imageProvider,
-    double size,
+  Widget renderShape({
+    String? txt,
+    ImageProvider? imageProvider,
+    double? size,
   }) {
-    return CircleAvatar(
-      radius: size,
-      backgroundColor: Palette.bgColor,
-      backgroundImage: imageProvider,
-      child: txt != null
-          ? Center(
+    if (imageProvider == null) {
+      var random = Random();
+      var colorIdx = random.nextInt(FrappePalette.colors.length);
+      var backgroundColor = FrappePalette.colors[colorIdx][100];
+      var textColor = FrappePalette.colors[colorIdx][600];
+
+      if (shape == ImageShape.circle) {
+        return CircleAvatar(
+          radius: size,
+          backgroundColor: backgroundColor,
+          child: txt != null
+              ? Center(
+                  child: Text(
+                    txt,
+                    style: TextStyle(
+                      fontSize: 12,
+                      color: textColor,
+                    ),
+                  ),
+                )
+              : null,
+        );
+      } else {
+        return ClipRRect(
+          borderRadius: BorderRadius.circular(
+            12,
+          ),
+          child: Container(
+            height: size,
+            width: size,
+            color: backgroundColor,
+            child: Center(
               child: Text(
-                txt,
-                style: TextStyle(fontSize: 12),
+                txt!,
+                style: TextStyle(
+                  fontSize: 12,
+                  color: textColor,
+                ),
               ),
-            )
-          : null,
-    );
+            ),
+          ),
+        );
+      }
+    } else {
+      if (shape == ImageShape.circle) {
+        return CircleAvatar(
+          radius: size,
+          backgroundImage: imageProvider,
+        );
+      } else {
+        return Image(
+          height: size,
+          width: size,
+          image: imageProvider,
+        );
+      }
+    }
   }
 
-  Future<Widget> getAvatar(String uid) async {
+  Widget getAvatar(
+    String? uid,
+  ) {
     if (uid == null) {
       return Container();
     }
-    var allUsers = await OfflineStorage.getItem('allUsers');
+    var allUsers = OfflineStorage.getItem('allUsers');
     allUsers = allUsers["data"];
     if (allUsers != null) {
       var user = allUsers[uid];
@@ -55,36 +102,36 @@ class UserAvatar extends StatelessWidget {
         return CachedNetworkImage(
           imageUrl: imageUrl,
           httpHeaders: {
-            HttpHeaders.cookieHeader: DioHelper.cookies,
+            // HttpHeaders.cookieHeader: DioHelper.cookies!,
           },
-          imageBuilder: (context, imageProvider) => UserAvatar.renderShape(
+          imageBuilder: (context, imageProvider) => renderShape(
             imageProvider: imageProvider,
             size: size,
           ),
-          placeholder: (context, url) => UserAvatar.renderShape(
+          placeholder: (context, url) => renderShape(
             txt: getInitials(
               user["full_name"],
             ),
             size: size,
           ),
-          errorWidget: (context, url, error) => UserAvatar.renderShape(
+          errorWidget: (context, url, error) => renderShape(
             txt: '',
             size: size,
           ),
         );
       } else if (user == null) {
-        return UserAvatar.renderShape(
+        return renderShape(
           txt: uid[0].toUpperCase(),
           size: size,
         );
       } else {
-        return UserAvatar.renderShape(
+        return renderShape(
           txt: getInitials(user["full_name"]),
           size: size,
         );
       }
     } else {
-      return UserAvatar.renderShape(
+      return renderShape(
         txt: uid[0].toUpperCase(),
         size: size,
       );
@@ -93,17 +140,6 @@ class UserAvatar extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return FutureBuilder(
-      future: getAvatar(uid),
-      builder: (context, snapshot) {
-        if (snapshot.hasData) {
-          return snapshot.data;
-        } else if (snapshot.hasError) {
-          return Text(snapshot.error);
-        } else {
-          return Container();
-        }
-      },
-    );
+    return getAvatar(uid);
   }
 }
