@@ -5,7 +5,7 @@ import 'locator.config.dart';
 
 import '../services/api/api.dart';
 import '../services/api/dio_api.dart';
-import '../services/api/fake_api.dart';
+// import '../services/api/fake_api.dart';
 
 final locator = GetIt.instance;
 
@@ -13,8 +13,11 @@ const bool USE_FAKE_IMPLEMENTATION = false;
 
 @injectableInit
 void setupLocator() {
+  // locator.registerLazySingleton<Api>(
+  //   () => USE_FAKE_IMPLEMENTATION ? FakeApi() : DioApi(),
+  // );
   locator.registerLazySingleton<Api>(
-    () => USE_FAKE_IMPLEMENTATION ? FakeApi() : DioApi(),
+    () => DioApi(),
   );
   $initGetIt(locator);
 }
