@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
+import 'package:form_builder_validators/form_builder_validators.dart';
 
 import '../../model/doctype_response.dart';
 
 class ControlInput {
-  String Function(dynamic) Function(BuildContext, {String errorText})
+  String? Function(dynamic) Function(BuildContext, {String errorText})?
       setMandatory(
     DoctypeField doctypeField,
   ) {
