@@ -1,17 +1,22 @@
+// @dart=2.9
+import 'dart:convert';
 import 'dart:io';
 
-import 'package:connectivity/connectivity.dart';
 import 'package:device_info/device_info.dart';
-import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_downloader/flutter_downloader.dart';
+import 'package:flutter_js/flutter_js.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
-import 'package:frappe_app/app/router.gr.dart';
+import 'package:frappe_app/model/common.dart';
 import 'package:frappe_app/model/offline_storage.dart';
 import 'package:frappe_app/services/storage_service.dart';
+import 'package:frappe_app/utils/frappe_alert.dart';
+import 'package:frappe_app/utils/navigation_helper.dart';
+import 'package:frappe_app/views/login/login_view.dart';
+import 'package:frappe_app/widgets/frappe_button.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:shared_preferences/shared_preferences.dart';
+import 'package:permission_handler/permission_handler.dart';
 
 import '../model/config.dart';
 import '../model/doctype_response.dart';
@@ -21,44 +26,10 @@ import '../views/no_internet.dart';
 
 import 'http.dart';
 import '../main.dart';
-import '../form/controls/control.dart';
 import '../app/locator.dart';
-import '../config/palette.dart';
-import '../services/navigation_service.dart';
 
 import '../utils/dio_helper.dart';
 import '../utils/enums.dart';
-
-import '../widgets/section.dart';
-import '../widgets/custom_expansion_tile.dart';
-
-// TODO
-Widget buildDecoratedWidget(Widget fieldWidget, bool withLabel,
-    [String label = ""]) {
-  if (withLabel) {
-    return Padding(
-      padding: Palette.fieldPadding,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          Padding(
-            padding: Palette.labelPadding,
-            child: Text(
-              label ?? "",
-              style: Palette.secondaryTxtStyle,
-            ),
-          ),
-          fieldWidget
-        ],
-      ),
-    );
-  } else {
-    return Padding(
-      padding: Palette.fieldPadding,
-      child: fieldWidget,
-    );
-  }
-}
 
 getDownloadPath() async {
   // TODO
@@ -78,7 +49,7 @@ downloadFile(String fileUrl, String downloadPath) async {
 
   await FlutterDownloader.enqueue(
     headers: {
-      HttpHeaders.cookieHeader: await DioHelper.getCookies(),
+      HttpHeaders.cookieHeader: DioHelper.cookies,
     },
     url: absoluteUrl,
     savedDir: downloadPath,
@@ -90,25 +61,14 @@ downloadFile(String fileUrl, String downloadPath) async {
 }
 
 Future<bool> _checkPermission() async {
-  // TODO
-  return true;
-  // if (Platform.isAndroid) {
-  //   PermissionStatus permission = await PermissionHandler()
-  //       .checkPermissionStatus(PermissionGroup.storage);
-  //   if (permission != PermissionStatus.granted) {
-  //     Map<PermissionGroup, PermissionStatus> permissions =
-  //         await PermissionHandler()
-  //             .requestPermissions([PermissionGroup.storage]);
-  //     if (permissions[PermissionGroup.storage] == PermissionStatus.granted) {
-  //       return true;
-  //     }
-  //   } else {
-  //     return true;
-  //   }
-  // } else {
-  //   return true;
-  // }
-  // return false;
+  if (Platform.isAndroid) {
+    if (await Permission.storage.request().isGranted) {
+      return true;
+    }
+  } else {
+    return true;
+  }
+  return false;
 }
 
 String toTitleCase(String str) {
@@ -121,211 +81,8 @@ String toTitleCase(String str) {
       .replaceAll(RegExp(r'(_|-)+'), ' ');
 }
 
-void showSnackBar(String txt, context) {
-  Scaffold.of(context).showSnackBar(
-    SnackBar(
-      content: Text(txt),
-    ),
-  );
-}
-
-List<Widget> generateLayout({
-  @required List<DoctypeField> fields,
-  @required ViewType viewType,
-  Map doc,
-  bool editMode = true,
-  bool withLabel = true,
-  Function onChanged,
-}) {
-  List<Widget> collapsibles = [];
-  List<Widget> sections = [];
-  List<Widget> widgets = [];
-
-  List<String> collapsibleLabels = [];
-  List<String> sectionLabels = [];
-
-  bool isCollapsible = false;
-  bool isSection = false;
-
-  int cIdx = 0;
-  int sIdx = 0;
-
-  fields.forEach((field) {
-    var val = doc != null
-        ? doc[field.fieldname] ?? field.defaultValue
-        : field.defaultValue;
-
-    if (val == '__user') {
-      val = Config().userId;
-    }
-
-    if (val is List) {
-      if (val.isEmpty) {
-        val = null;
-      }
-    }
-
-    if (field.fieldtype == "Section Break") {
-      if (sections.length > 0) {
-        var sectionVisibility = sections.any((element) {
-          if (element is Visibility) {
-            return element.visible == true;
-          } else {
-            return true;
-          }
-        });
-
-        widgets.add(
-          Visibility(
-            visible: sectionVisibility,
-            child: sectionLabels[sIdx] != ''
-                ? ListTileTheme(
-                    contentPadding: EdgeInsets.all(0),
-                    child: CustomExpansionTile(
-                      maintainState: true,
-                      initiallyExpanded: true,
-                      title: Text(
-                        sectionLabels[sIdx].toUpperCase(),
-                        style: Palette.secondaryTxtStyle,
-                      ),
-                      children: [...sections],
-                    ),
-                  )
-                : Section(
-                    title: sectionLabels[sIdx],
-                    children: [...sections],
-                  ),
-          ),
-        );
-
-        sIdx += 1;
-        sections.clear();
-      } else if (collapsibles.length > 0) {
-        var sectionVisibility = collapsibles.any((element) {
-          if (element is Visibility) {
-            return element.visible == true;
-          } else {
-            return true;
-          }
-        });
-        widgets.add(
-          Visibility(
-            visible: sectionVisibility,
-            child: ListTileTheme(
-              contentPadding: EdgeInsets.all(0),
-              child: CustomExpansionTile(
-                maintainState: true,
-                title: Text(
-                  collapsibleLabels[cIdx].toUpperCase(),
-                  style: Palette.secondaryTxtStyle,
-                ),
-                children: [...collapsibles],
-              ),
-            ),
-          ),
-        );
-        cIdx += 1;
-        collapsibles.clear();
-      }
-
-      if (field.collapsible == 1) {
-        isSection = false;
-        isCollapsible = true;
-        collapsibleLabels.add(field.label);
-      } else {
-        isCollapsible = false;
-        isSection = true;
-        sectionLabels.add(field.label != null ? field.label.toUpperCase() : '');
-      }
-    } else if (isCollapsible) {
-      if (viewType == ViewType.form) {
-        collapsibles.add(
-          Visibility(
-            visible: editMode ? true : val != null && val != '',
-            child: makeControl(
-              field: field,
-              value: val,
-              doc: doc,
-              withLabel: withLabel,
-              editMode: editMode,
-              onChanged: onChanged,
-            ),
-          ),
-        );
-      } else {
-        collapsibles.add(
-          makeControl(
-            doc: doc,
-            field: field,
-            value: val,
-            withLabel: withLabel,
-            editMode: editMode,
-            onChanged: onChanged,
-          ),
-        );
-      }
-    } else if (isSection) {
-      if (viewType == ViewType.form) {
-        sections.add(
-          Visibility(
-            visible: editMode ? true : val != null && val != '',
-            child: makeControl(
-              doc: doc,
-              field: field,
-              value: val,
-              withLabel: withLabel,
-              editMode: editMode,
-              onChanged: onChanged,
-            ),
-          ),
-        );
-      } else {
-        sections.add(
-          makeControl(
-            field: field,
-            doc: doc,
-            value: val,
-            withLabel: withLabel,
-            editMode: editMode,
-            onChanged: onChanged,
-          ),
-        );
-      }
-    } else {
-      if (viewType == ViewType.form) {
-        widgets.add(
-          Visibility(
-            visible: editMode ? true : val != null && val != '',
-            child: makeControl(
-              doc: doc,
-              field: field,
-              value: val,
-              withLabel: withLabel,
-              editMode: editMode,
-              onChanged: onChanged,
-            ),
-          ),
-        );
-      } else {
-        widgets.add(
-          makeControl(
-            field: field,
-            value: val,
-            doc: doc,
-            withLabel: withLabel,
-            editMode: editMode,
-            onChanged: onChanged,
-          ),
-        );
-      }
-    }
-  });
-
-  return widgets;
-}
-
 DateTime parseDate(val) {
-  if (val == null) {
+  if (val == null || val == "") {
     return null;
   } else if (val == "Today") {
     return DateTime.now();
@@ -406,25 +163,52 @@ getTitle(DoctypeDoc meta, Map doc) {
 }
 
 clearLoginInfo() async {
-  await DioHelper.getCookiePath()
-    ..delete(
+  var cookie = await DioHelper.getCookiePath();
+  if (Config().uri != null) {
+    cookie.delete(
       Config().uri,
     );
+  }
+
   Config.set('isLoggedIn', false);
 }
 
-handle403() async {
+handle403(BuildContext context) async {
   await clearLoginInfo();
-  locator<NavigationService>().clearAllAndNavigateTo(Routes.sessionExpired);
+
+  NavigationHelper.clearAllAndNavigateTo(
+    context: context,
+    page: Login(),
+  );
 }
 
-handleError(Response error, [bool hideAppBar = false]) {
+handleError({
+  @required ErrorResponse error,
+  @required BuildContext context,
+  Function onRetry,
+  bool hideAppBar = false,
+}) {
   if (error.statusCode == HttpStatus.forbidden) {
-    handle403();
+    handle403(context);
   } else if (error.statusCode == HttpStatus.serviceUnavailable) {
     return NoInternet(hideAppBar);
   } else {
-    return Scaffold(appBar: AppBar(), body: Text("${error.statusMessage}"));
+    return Scaffold(
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Text("${error.statusMessage}"),
+            if (onRetry != null)
+              FrappeFlatButton(
+                onPressed: () => onRetry(),
+                buttonType: ButtonType.primary,
+                title: "Retry",
+              ),
+          ],
+        ),
+      ),
+    );
   }
 }
 
@@ -468,25 +252,6 @@ Future<int> getActiveNotifications() async {
   return activeNotifications.length;
 }
 
-showErrorDialog(e, BuildContext context) {
-  showDialog(
-    context: context,
-    builder: (context) {
-      return AlertDialog(
-        title: Text(e.statusMessage),
-        actions: <Widget>[
-          FlatButton(
-            child: Text('Ok'),
-            onPressed: () async {
-              locator<NavigationService>().pop();
-            },
-          ),
-        ],
-      );
-    },
-  );
-}
-
 Map extractChangedValues(Map original, Map updated) {
   var changedValues = {};
   for (var key in updated.keys) {
@@ -495,27 +260,6 @@ Map extractChangedValues(Map original, Map updated) {
     }
   }
   return changedValues;
-}
-
-requestIOSLocationAuthorization(connectivity) async {
-  try {
-    if (Platform.isIOS) {
-      LocationAuthorizationStatus status =
-          await connectivity.getLocationServiceAuthorization();
-      if (status == LocationAuthorizationStatus.notDetermined) {
-        status = await connectivity.requestLocationServiceAuthorization();
-      } else {
-        return {
-          "access": true,
-        };
-      }
-    }
-  } on PlatformException catch (e) {
-    print(e.toString());
-    return {
-      "access": false,
-    };
-  }
 }
 
 Future<bool> verifyOnline() async {
@@ -546,28 +290,19 @@ getLinkFields(String doctype) async {
   return linkFieldDoctypes;
 }
 
-putSharedPrefValue(String key, bool value) async {
-  var _prefs = await SharedPreferences.getInstance();
-  await _prefs.setBool(key, value);
-}
-
-Future<bool> getSharedPrefValue(String key) async {
-  var _prefs = await SharedPreferences.getInstance();
-  await _prefs.reload();
-  return _prefs.getBool(key);
-}
-
 resetValues() async {
-  await putSharedPrefValue("backgroundTask", false);
-  await putSharedPrefValue("storeApiResponse", true);
+  await locator<StorageService>()
+      .putSharedPrefBoolValue("backgroundTask", false);
+  await locator<StorageService>()
+      .putSharedPrefBoolValue("storeApiResponse", true);
 }
 
 initDb() async {
-  await locator<StorageService>().initStorage();
+  await locator<StorageService>().initHiveStorage();
 
-  await locator<StorageService>().initBox('queue');
-  await locator<StorageService>().initBox('offline');
-  await locator<StorageService>().initBox('config');
+  await locator<StorageService>().initHiveBox('queue');
+  await locator<StorageService>().initHiveBox('offline');
+  await locator<StorageService>().initHiveBox('config');
 }
 
 initLocalNotifications() async {
@@ -590,21 +325,63 @@ initAwesomeItems() async {
   var moduleDoctypesMapping = {};
 
   for (var item in deskSidebarItems.message) {
-    var desktopPage = await locator<Api>().getDesktopPage(item.name);
+    String module;
+    if (item.content != null) {
+      module = jsonEncode({
+        "name": item.name,
+        "title": item.name,
+        "content": item.content,
+      });
+    } else {
+      module = item.name;
+    }
 
+    var desktopPage = await locator<Api>().getDesktopPage(module);
+
+    var doctypes = [];
     desktopPage.message.cards.items.forEach(
       (item) {
-        var doctypes = [];
         item.links.forEach(
           (link) {
             doctypes.add(link.label);
           },
         );
-
-        moduleDoctypesMapping[item.label] = doctypes;
       },
     );
+    moduleDoctypesMapping[item.label] = doctypes;
   }
 
   OfflineStorage.putItem('awesomeItems', moduleDoctypesMapping);
+}
+
+noInternetAlert(
+  BuildContext context,
+) {
+  FrappeAlert.warnAlert(
+    title: "Connection Lost",
+    subtitle: "You are not connected to Internet. Retry after sometime.",
+    context: context,
+  );
+}
+
+executeJS({
+  @required String jsString,
+}) {
+  JavascriptRuntime flutterJs = getJavascriptRuntime();
+  try {
+    JsEvalResult jsResult = flutterJs.evaluate(jsString);
+    return jsResult.rawResult;
+  } on PlatformException catch (e) {
+    print('ERRO: ${e.details}');
+  }
+}
+
+String getServerMessage(String serverMsgs) {
+  var errorMsgs = json.decode(serverMsgs) as List;
+  var errorStr = '';
+  errorMsgs.forEach((errorMsg) {
+    errorStr += json.decode(errorMsg)["message"];
+  });
+
+  return errorStr;
 }
