@@ -7,12 +7,13 @@ import '../model/config.dart';
 
 initApiConfig() async {
   if (Config().baseUrl != null) {
-    await DioHelper.init(Config().baseUrl);
+    await DioHelper.init(Config().baseUrl!);
+    await DioHelper.initCookies();
   }
 }
 
 Future<void> cacheAllUsers() async {
-  var allUsers = await OfflineStorage.getItem('allUsers');
+  var allUsers = OfflineStorage.getItem('allUsers');
   allUsers = allUsers["data"];
   if (allUsers != null) {
     return;
@@ -27,20 +28,27 @@ Future<void> cacheAllUsers() async {
       ["User", "enabled", "=", 1]
     ];
 
-    var meta = await locator<Api>().getDoctype('User');
+    try {
+      var meta = await locator<Api>().getDoctype('User');
 
-    var res = await locator<Api>().fetchList(
-      fieldnames: fieldNames,
-      doctype: 'User',
-      filters: filters,
-      meta: meta.docs[0],
-    );
+      var res = await locator<Api>().fetchList(
+        fieldnames: fieldNames,
+        doctype: 'User',
+        orderBy: '`tabUser`.`modified` desc',
+        filters: filters,
+        meta: meta.docs[0],
+      );
 
-    var usr = {};
-    res.forEach((element) {
-      usr[element["name"]] = element;
-    });
-    OfflineStorage.putItem('allUsers', usr);
+      var usr = {};
+      res.forEach(
+        (element) {
+          usr[element["name"]] = element;
+        },
+      );
+      OfflineStorage.putItem('allUsers', usr);
+    } catch (e) {
+      throw e;
+    }
   }
 }
 
