@@ -5,12 +5,12 @@ import 'package:provider/provider.dart';
 import 'base_viewmodel.dart';
 
 class BaseView<T extends BaseViewModel> extends StatefulWidget {
-  final Widget Function(BuildContext context, T model, Widget child) builder;
-  final Function(T) onModelReady;
-  final Function(T) onModelClose;
+  final Widget Function(BuildContext context, T model, Widget? child) builder;
+  final Function(T)? onModelReady;
+  final Function(T)? onModelClose;
 
   BaseView({
-    this.builder,
+    required this.builder,
     this.onModelReady,
     this.onModelClose,
   });
@@ -25,7 +25,7 @@ class _BaseViewState<T extends BaseViewModel> extends State<BaseView<T>> {
   @override
   void initState() {
     if (widget.onModelReady != null) {
-      widget.onModelReady(model);
+      widget.onModelReady!(model);
     }
     super.initState();
   }
@@ -33,7 +33,7 @@ class _BaseViewState<T extends BaseViewModel> extends State<BaseView<T>> {
   @override
   void dispose() {
     if (widget.onModelClose != null) {
-      widget.onModelClose(model);
+      widget.onModelClose!(model);
     }
     super.dispose();
   }
